update vector integrations section
