+++ apps/www/data/products/vector/pageData.tsx
@@ -73,7 +73,6 @@ export default {
   },
   integrations: {
     title: 'Leverage the tools you love',
-    image: IntegrationsImage,
   },
   useCasesSection: {
     title: (
